Several improvements.
