+++ python/aristotle-metadata-registry/aristotle_mdr/signals.py
@@ -1,16 +1,19 @@
+from django.core.exceptions import ImproperlyConfigured
 from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
-from django.db import transaction
-from django.db.models.signals import m2m_changed, post_save, pre_delete
+from django.db.models.signals import m2m_changed, pre_save, post_save, pre_delete
 from haystack import signals as haystack_signals
+from reversion.signals import post_revision_commit
+from django.dispatch import receiver
 
-from aristotle_mdr.contrib.async_signals.utils import clean_signal
+from aristotle_mdr.contrib.async_signals.utils import clean_signal, fire
 from aristotle_bg_workers.tasks import update_search_index, delete_search_index
 from aristotle_bg_workers.utils import run_task_on_commit
 from aristotle_mdr.utils import fetch_metadata_apps
+from aristotle_mdr.models import _concept, DiscussionComment, DiscussionPost, Status, PossumProfile, SupersedeRelationship
 
 # Don't import aristotle_mdr.models directly, only pull in whats required,
-#  otherwise Haystack gets into a circular dependancy.
+#  otherwise Haystack gets into a circular dependency.
 
 import logging
 
@@ -143,3 +146,91 @@ class AristotleSignalProcessor(haystack_signals.BaseSignalProcessor):
                 },
             ]
             delete_search_index.delay(*args, **message)
+
+
+@receiver(pre_save)
+def check_concept_app_label(sender, instance, **kwargs):
+    if not issubclass(sender, _concept):
+        return
+    if instance._meta.app_label not in fetch_metadata_apps():
+        raise ImproperlyConfigured(
+            (
+                "Trying to save item <{instance_name}> when app_label <{app_label}> "
+                "is not enabled. Metadata apps at this point were <{metadata_apps}>."
+            ).format(
+                app_label=instance._meta.app_label,
+                instance_name=instance.name,
+                metadata_apps=str(fetch_metadata_apps())
+            )
+        )
+
+
+@receiver(pre_save)
+def update_org_to_match_workgroup(sender, instance, **kwargs):
+    """Enforces integrity between Stewardship Organisation and Workgroup"""
+    if not issubclass(sender, _concept):
+        return
+    if instance.workgroup is not None:
+        instance.stewardship_organisation = instance.workgroup.stewardship_organisation
+
+
+@receiver(post_save, sender=DiscussionComment)
+def new_comment_created(sender, instance, **kwargs):
+    comment = instance
+    post = comment.post
+    if kwargs.get('raw'):
+        # Don't run during loaddata
+        return
+    if not kwargs['created']:
+        return  # We don't need to notify a topic poster of an edit.
+    if comment.author == post.author:
+        return  # We don't need to tell someone they replied to themselves
+    fire("notification_events.new_comment_created", obj=comment, **kwargs)
+
+
+@receiver(post_save, sender=DiscussionPost)
+def new_post_created(sender, instance, **kwargs):
+    # Don't pass the instance to the JSONEncoder
+    post = instance
+    if kwargs.get('raw'):
+        # Don't run during loaddata
+        return
+    if not kwargs['created']:
+        return  # We don't need to notify a topic poster of an edit.
+    fire("notification_events.new_post_created", obj=post, **kwargs)
+
+
+@receiver(post_save, sender=Status)
+def states_changed(sender, instance, *args, **kwargs):
+    fire("notification_events.status_changed", obj=instance, **kwargs)
+
+
+@receiver(post_save, sender=SupersedeRelationship)
+def new_superseded_relation(sender, instance, *args, **kwargs):
+    if kwargs.get('created'):
+        fire("notification_events.item_superseded", obj=instance, **kwargs)
+
+
+@receiver(post_save, sender=settings.AUTH_USER_MODEL)
+def create_user_profile(sender, instance, created, **kwargs):
+    if created:
+        PossumProfile.objects.get_or_create(user=instance)
+
+
+@receiver(post_revision_commit)
+def concept_saved(sender, **kwargs):
+
+    revision = kwargs.pop('revision')
+    versions = kwargs.pop('versions')
+    version = versions[0]
+
+    if not issubclass(version._model, _concept):
+        return
+
+    user_id = revision.user_id
+
+    if user_id is not None:
+
+        # If the concept saved was not triggered by a superseding action:
+        # if not ('modified' in changed_fields and len(changed_fields) == 1):
+        fire("notification_events.create_notifications_for_saved_concept", obj=version, user_id=user_id, **kwargs)
